# Conflicts:
#	CHANGELOG.md
#	src/Execution/Arguments/UpsertModel.php

+++ tests/database/migrations/2018_09_30_000007_create_testbench_taggables_table.php
@@ -9,11 +9,9 @@ final class CreateTestbenchTaggablesTable extends Migration
     public function up(): void
     {
         Schema::create('taggables', function (Blueprint $table): void {
-            $table->unsignedInteger('tag_id');
-
-            $table->unsignedInteger('taggable_id');
+            $table->unsignedBigInteger('tag_id');
+            $table->unsignedBigInteger('taggable_id');
             $table->string('taggable_type');
-
             $table->timestamps();
         });
     }
